Run script for instantiated template
This makes it possible to automatically instanciate and run the script template
and should help with testing.

Currently there is still some problem with parameter passing - if the run.sh
script has more than one argument, sbt only recognizes the first one of those as
a parameter to the csRun command. -> To be fixed